Separate decoder and ignore urlsession for coverage
